+++ localflavor/in_/models.py
@@ -1,10 +1,11 @@
-from django.db.models import CharField
+from django.db import models
 from django.utils.translation import gettext_lazy as _
 
 from .in_states import STATE_CHOICES
+from .forms import INPANCardNumberFormField
+from .validators import INPANCardNumberValidator
 
-
-class INStateField(CharField):
+class INStateField(models.CharField):
     """
     A model field that stores the two-letter Indian state abbreviation in the database.
 
@@ -22,3 +23,22 @@ class INStateField(CharField):
         name, path, args, kwargs = super().deconstruct()
         del kwargs['choices']
         return name, path, args, kwargs
+
+class INPANCardNumberField(models.CharField):
+    """
+    A model field that accepts indian PAN Card number.
+
+    Forms represent it as a ``forms.INPANCardNumberFormField`` field.
+
+    .. versionadded:: 4.0
+    """
+    description = _("PAN Card number field")
+    def __init__(self, *args, **kwargs):
+        kwargs['max_length'] = 10
+        super().__init__(*args, **kwargs)
+        self.validators.append(INPANCardNumberValidator())
+
+    def formfield(self, **kwargs):
+        defaults = {'form_class': INPANCardNumberFormField}
+        defaults.update(kwargs)
+        return super().formfield(**defaults)